Simple calculator application in C#
This is a simple calculator application I'm using as an example to practice writing in C#.  The solution is structured such that the backend (CalculatorModel) is separated from the user interfaces (ConsoleInterface and WPFInterface).  I wanted to make two different interfaces to the same codebase to ensure that I am writing code that is decoupled from the user interface.